Refactor the names of all React components
