+++ .gitignore
@@ -1,2 +1,3 @@
 
+build
 node_modules
