+++ package.json
@@ -3,6 +3,7 @@
   "description": "Record all your expenses!",
   "version": "1.0.0",
   "dependencies": {
+    "serve": "^14.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1"
