+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Expenses Recorder",
+  "name": "Expenses Recorder using React.js",
   "icons": [
     {
       "src": "favicon.ico",
